+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Support for field values of the type `double`, `ulong`, `short` and `ushort`
+
 ## 3.1.0 - 2018-08-29
 
 - Write request body to when getting exception while sending